refactor: move updateDevicesStatus from DevicesService to PqcGatewayService
